Add extra check

+++ bootstrap/setup.sh
@@ -52,7 +52,7 @@ if [ -f /etc/alpine-release ]; then
     apk add -q $BASEDIR/files/glibc-2.26-r0.apk || apk add libc6-compat || true
 
     # symlink musl as a last resort
-    if [ ! -f /lib64/ld-linux-x86-64.so.2 ]; then
+    if [ ! -f /lib64/ld-linux-x86-64.so.2 ] && [ -f /lib/libc.musl-x86_64.so.1 ]; then
       mkdir -p /lib64
       ln -s /lib/libc.musl-x86_64.so.1 /lib64/ld-linux-x86-64.so.2
     fi
